requirements.txt
aniso8601==9.0.1
bidict==0.21.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
gevent==21.8.0
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
pygame==2.0.1
python-engineio==4.2.1
python-socketio==5.4.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
urllib3==1.26.6
websocket-client==1.2.1
websockets==9.1
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
